+++ pkg/storage/chunk/azure/blob_store_metrics.go
@@ -1,0 +1,37 @@
+package azure
+
+import "github.com/prometheus/client_golang/prometheus"
+
+type BlobStorageMetrics struct {
+	requestDuration  *prometheus.HistogramVec
+	egressBytesTotal prometheus.Counter
+}
+
+// NewBlobStorageMetrics creates the blob storage metrics struct and registers all of it's metrics.
+func NewBlobStorageMetrics() BlobStorageMetrics {
+	b := BlobStorageMetrics{
+		requestDuration: prometheus.NewHistogramVec(prometheus.HistogramOpts{
+			Namespace: "loki",
+			Name:      "azure_blob_request_duration_seconds",
+			Help:      "Time spent doing azure blob requests.",
+			// Latency seems to range from a few ms to a few secs and is
+			// important. So use 6 buckets from 5ms to 5s.
+			Buckets: prometheus.ExponentialBuckets(0.005, 4, 6),
+		}, []string{"operation", "status_code"}),
+		egressBytesTotal: prometheus.NewCounter(prometheus.CounterOpts{
+			Namespace: "loki",
+			Name:      "azure_blob_egress_bytes_total",
+			Help:      "Total bytes downloaded from Azure Blob Storage.",
+		}),
+	}
+	prometheus.MustRegister(b.requestDuration)
+	prometheus.MustRegister(b.egressBytesTotal)
+	return b
+}
+
+// Unregister unregisters the blob storage metrics with the prometheus default registerer, useful for tests
+// where we frequently need to create multiple instances of the metrics struct, but not globally.
+func (bm *BlobStorageMetrics) Unregister() {
+	prometheus.Unregister(bm.requestDuration)
+	prometheus.Unregister(bm.egressBytesTotal)
+}